Split Antfarm artifact workflow components
