+++ ts/apps/antfarm/src/pages/TableDetailsPage.tsx
@@ -50,9 +50,6 @@ import {
 } from "@antfly/design-system";
 import type {
   Table as AntflyTable,
-  DocumentArtifactManifest,
-  DocumentArtifactReprocessJob,
-  DocumentArtifactTableReprocessResponse,
   IndexStatus,
   QueryRequest,
   QueryResult,
@@ -76,6 +73,8 @@ import QueryBuilder from "../components/querybuilder/QueryBuilder";
 import { QueryResultsList } from "../components/results";
 import SearchBoxBuilder from "../components/SearchBoxBuilder";
 import DocumentSchemasForm from "../components/schema-builder/DocumentSchemasForm";
+import { DocumentArtifactsPanel } from "../components/table/DocumentArtifactsPanel";
+import { TableReprocessPanel } from "../components/table/TableReprocessPanel";
 import {
   type BasicField,
   generateBasicFields,
@@ -158,38 +157,6 @@ function getTotalIndexed(index: IndexStatus) {
     return (index.status as { total_indexed?: number }).total_indexed ?? "N/A";
   }
   return "N/A";
-}
-
-function formatCount(value: number | null | undefined) {
-  return value === null || value === undefined ? "-" : Intl.NumberFormat().format(value);
-}
-
-function parseOptionalLimit(value: string) {
-  const trimmed = value.trim();
-  if (!trimmed) return undefined;
-  const parsed = Number.parseInt(trimmed, 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : undefined;
-}
-
-function artifactHealthClass(artifact: DocumentArtifactManifest) {
-  if (artifact.last_error_code || artifact.unsupported_reason) return "af-status-badge-error";
-  if (artifact.merge_status && artifact.merge_status !== "complete")
-    return "af-status-badge-warning";
-  return "af-status-badge-success";
-}
-
-function artifactHealthLabel(artifact: DocumentArtifactManifest) {
-  if (artifact.last_error_code) return artifact.last_error_code;
-  if (artifact.unsupported_reason) return "unsupported";
-  return artifact.merge_status || "ready";
-}
-
-function jobStatusClass(job: DocumentArtifactReprocessJob) {
-  if (job.phase === "failed" || job.phase === "cancelled") return "af-status-badge-error";
-  if (job.phase === "running" || job.reprocess_status === "in_progress") {
-    return "af-status-badge-warning";
-  }
-  return "af-status-badge-success";
 }
 
 function buildIndexColumns(
@@ -279,24 +246,6 @@ const TableDetailsPage: React.FC<TableDetailsPageProps> = ({ currentSection = "o
   const [semanticQuery, setSemanticQuery] = useState(JSON.stringify({}, null, 2));
   const [selectedFields, setSelectedFields] = useState<string[]>([]);
   const [includeProfile, setIncludeProfile] = useState(true);
-  const [artifactDocumentKey, setArtifactDocumentKey] = useState("");
-  const [artifactManifests, setArtifactManifests] = useState<DocumentArtifactManifest[]>([]);
-  const [selectedArtifactName, setSelectedArtifactName] = useState("");
-  const [selectedArtifactDetail, setSelectedArtifactDetail] =
-    useState<DocumentArtifactManifest | null>(null);
-  const [artifactLoading, setArtifactLoading] = useState(false);
-  const [artifactMessage, setArtifactMessage] = useState<string | null>(null);
-  const [artifactError, setArtifactError] = useState<string | null>(null);
-  const [reprocessArtifactName, setReprocessArtifactName] = useState("document_units_v1");
-  const [reprocessFromKey, setReprocessFromKey] = useState("");
-  const [reprocessToKey, setReprocessToKey] = useState("");
-  const [reprocessLimit, setReprocessLimit] = useState("100");
-  const [reprocessJob, setReprocessJob] = useState<DocumentArtifactReprocessJob | null>(null);
-  const [reprocessPass, setReprocessPass] = useState<DocumentArtifactTableReprocessResponse | null>(
-    null
-  );
-  const [reprocessLoading, setReprocessLoading] = useState(false);
-  const [reprocessError, setReprocessError] = useState<string | null>(null);
 
   // Derive search modes from input content instead of toggles
   const hasSemanticQuery = query.trim().length > 0 && queryIndexes.length > 0;
@@ -464,174 +413,6 @@ const TableDetailsPage: React.FC<TableDetailsPageProps> = ({ currentSection = "o
       setDocumentCount(null);
     }
   }, [storageStatus?.empty, tableName]);
-
-  const loadArtifactDetail = useCallback(
-    async (artifactName: string) => {
-      if (!tableName || !artifactDocumentKey.trim() || !artifactName) return;
-      setArtifactLoading(true);
-      setArtifactError(null);
-      try {
-        const detail = await api.tables.artifacts.get(
-          tableName,
-          artifactDocumentKey.trim(),
-          artifactName,
-          "raw"
-        );
-        setSelectedArtifactDetail(detail ?? null);
-      } catch (e) {
-        setArtifactError(`Failed to load artifact ${artifactName}.`);
-        console.error(e);
-      } finally {
-        setArtifactLoading(false);
-      }
-    },
-    [artifactDocumentKey, tableName]
-  );
-
-  const loadArtifactManifests = useCallback(async () => {
-    if (!tableName || !artifactDocumentKey.trim()) return;
-    setArtifactLoading(true);
-    setArtifactError(null);
-    setArtifactMessage(null);
-    try {
-      const response = await api.tables.artifacts.list(tableName, artifactDocumentKey.trim());
-      const artifacts = response?.artifacts ?? [];
-      setArtifactManifests(artifacts);
-      const nextSelected =
-        artifacts.find((artifact) => artifact.artifact_name === selectedArtifactName)
-          ?.artifact_name ??
-        artifacts[0]?.artifact_name ??
-        "";
-      setSelectedArtifactName(nextSelected);
-      if (nextSelected) {
-        const detail = await api.tables.artifacts.get(
-          tableName,
-          artifactDocumentKey.trim(),
-          nextSelected,
-          "raw"
-        );
-        setSelectedArtifactDetail(detail ?? null);
-      } else {
-        setSelectedArtifactDetail(null);
-      }
-    } catch (e) {
-      setArtifactManifests([]);
-      setSelectedArtifactDetail(null);
-      setArtifactError(`No artifact manifests found for ${artifactDocumentKey.trim()}.`);
-      console.error(e);
-    } finally {
-      setArtifactLoading(false);
-    }
-  }, [artifactDocumentKey, selectedArtifactName, tableName]);
-
-  const handleSelectArtifact = async (artifactName: string) => {
-    setSelectedArtifactName(artifactName);
-    await loadArtifactDetail(artifactName);
-  };
-
-  const handleReprocessDocumentArtifact = async () => {
-    if (!tableName || !artifactDocumentKey.trim() || !selectedArtifactName) return;
-    setArtifactLoading(true);
-    setArtifactError(null);
-    setArtifactMessage(null);
-    try {
-      await api.tables.artifacts.reprocessDocument(
-        tableName,
-        artifactDocumentKey.trim(),
-        selectedArtifactName
-      );
-      setArtifactMessage(`Reprocess accepted for ${selectedArtifactName}.`);
-      await loadArtifactDetail(selectedArtifactName);
-    } catch (e) {
-      setArtifactError(`Failed to reprocess ${selectedArtifactName}.`);
-      console.error(e);
-    } finally {
-      setArtifactLoading(false);
-    }
-  };
-
-  const reprocessRequest = () => ({
-    ...(reprocessFromKey.trim() ? { from_key: reprocessFromKey.trim() } : {}),
-    ...(reprocessToKey.trim() ? { to_key: reprocessToKey.trim() } : {}),
-    ...(parseOptionalLimit(reprocessLimit) ? { limit: parseOptionalLimit(reprocessLimit) } : {}),
-  });
-
-  const handleRunReprocessPass = async () => {
-    if (!tableName || !reprocessArtifactName.trim()) return;
-    setReprocessLoading(true);
-    setReprocessError(null);
-    try {
-      const response = await api.tables.artifacts.reprocessRange(
-        tableName,
-        reprocessArtifactName.trim(),
-        reprocessRequest()
-      );
-      setReprocessPass(response ?? null);
-    } catch (e) {
-      setReprocessError(`Failed to run reprocess pass for ${reprocessArtifactName.trim()}.`);
-      console.error(e);
-    } finally {
-      setReprocessLoading(false);
-    }
-  };
-
-  const handleStartReprocessJob = async () => {
-    if (!tableName || !reprocessArtifactName.trim()) return;
-    setReprocessLoading(true);
-    setReprocessError(null);
-    try {
-      const response = await api.tables.artifacts.startReprocessJob(
-        tableName,
-        reprocessArtifactName.trim(),
-        { ...reprocessRequest(), advance: true }
-      );
-      setReprocessJob(response ?? null);
-      setReprocessPass(null);
-    } catch (e) {
-      setReprocessError(`Failed to start reprocess job for ${reprocessArtifactName.trim()}.`);
-      console.error(e);
-    } finally {
-      setReprocessLoading(false);
-    }
-  };
-
-  const handleAdvanceReprocessJob = async () => {
-    if (!tableName || !reprocessJob) return;
-    setReprocessLoading(true);
-    setReprocessError(null);
-    try {
-      const response = await api.tables.artifacts.advanceReprocessJob(
-        tableName,
-        reprocessJob.artifact_name,
-        String(reprocessJob.job_id)
-      );
-      setReprocessJob(response ?? null);
-    } catch (e) {
-      setReprocessError(`Failed to advance job ${reprocessJob.job_id}.`);
-      console.error(e);
-    } finally {
-      setReprocessLoading(false);
-    }
-  };
-
-  const handleCancelReprocessJob = async () => {
-    if (!tableName || !reprocessJob) return;
-    setReprocessLoading(true);
-    setReprocessError(null);
-    try {
-      const response = await api.tables.artifacts.cancelReprocessJob(
-        tableName,
-        reprocessJob.artifact_name,
-        String(reprocessJob.job_id)
-      );
-      setReprocessJob(response ?? null);
-    } catch (e) {
-      setReprocessError(`Failed to cancel job ${reprocessJob.job_id}.`);
-      console.error(e);
-    } finally {
-      setReprocessLoading(false);
-    }
-  };
 
   useEffect(() => {
     fetchIndexes();
@@ -1498,336 +1279,12 @@ const TableDetailsPage: React.FC<TableDetailsPageProps> = ({ currentSection = "o
 
         {/* Artifacts Section */}
         {currentSection === "artifacts" && (
-          <div className="grid gap-4 xl:grid-cols-[360px_1fr]">
-            <Card>
-              <CardHeader>
-                <CardTitle>Document Artifacts</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="space-y-2">
-                  <Label htmlFor="artifact-document-key">Document key</Label>
-                  <div className="flex gap-2">
-                    <Input
-                      id="artifact-document-key"
-                      value={artifactDocumentKey}
-                      onChange={(event) => setArtifactDocumentKey(event.target.value)}
-                      onKeyDown={(event) => {
-                        if (event.key === "Enter") {
-                          event.preventDefault();
-                          loadArtifactManifests();
-                        }
-                      }}
-                      placeholder="doc:a"
-                    />
-                    <Button onClick={loadArtifactManifests} disabled={artifactLoading}>
-                      Load
-                    </Button>
-                  </div>
-                </div>
-
-                {artifactError && (
-                  <Alert variant="destructive">
-                    <AlertDescription>{artifactError}</AlertDescription>
-                  </Alert>
-                )}
-                {artifactMessage && (
-                  <Alert>
-                    <AlertDescription>{artifactMessage}</AlertDescription>
-                  </Alert>
-                )}
-
-                <div className="space-y-2">
-                  <div className="text-sm font-medium">Manifests</div>
-                  {artifactManifests.length === 0 ? (
-                    <div className="rounded border border-dashed p-4 text-sm text-muted-foreground">
-                      Enter a document key to inspect derived artifacts.
-                    </div>
-                  ) : (
-                    <div className="space-y-2">
-                      {artifactManifests.map((artifact) => (
-                        <button
-                          key={artifact.artifact_name}
-                          type="button"
-                          className={`w-full rounded border p-3 text-left transition-colors hover:bg-muted ${
-                            selectedArtifactName === artifact.artifact_name
-                              ? "border-primary bg-muted"
-                              : "border-border"
-                          }`}
-                          onClick={() => handleSelectArtifact(artifact.artifact_name)}
-                        >
-                          <div className="flex items-center justify-between gap-2">
-                            <span className="font-medium">{artifact.artifact_name}</span>
-                            <Badge className={artifactHealthClass(artifact)}>
-                              {artifactHealthLabel(artifact)}
-                            </Badge>
-                          </div>
-                          <div className="mt-2 grid grid-cols-3 gap-2 text-xs text-muted-foreground">
-                            <span>{formatCount(artifact.unit_count)} units</span>
-                            <span>{formatCount(artifact.chunk_count)} chunks</span>
-                            <span>gen {artifact.generation}</span>
-                          </div>
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                </div>
-
-                <Button
-                  variant="outline"
-                  onClick={handleReprocessDocumentArtifact}
-                  disabled={artifactLoading || !selectedArtifactName}
-                  className="w-full"
-                >
-                  Reprocess selected artifact
-                </Button>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader>
-                <CardTitle>Artifact Detail</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                {selectedArtifactDetail ? (
-                  <>
-                    <div className="grid gap-3 md:grid-cols-4">
-                      <div>
-                        <div className="text-xs text-muted-foreground">Content Type</div>
-                        <div className="text-sm font-medium">
-                          {selectedArtifactDetail.content_type || "-"}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Route</div>
-                        <div className="text-sm font-medium">
-                          {selectedArtifactDetail.route_type || "-"}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Ranges</div>
-                        <div className="text-sm font-medium">
-                          {formatCount(selectedArtifactDetail.child_range_count)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Merge</div>
-                        <div className="text-sm font-medium">
-                          {selectedArtifactDetail.merge_status || "-"}
-                        </div>
-                      </div>
-                    </div>
-
-                    {(selectedArtifactDetail.last_error_message ||
-                      selectedArtifactDetail.unsupported_reason) && (
-                      <Alert variant="destructive">
-                        <AlertDescription>
-                          {selectedArtifactDetail.last_error_message ||
-                            selectedArtifactDetail.unsupported_reason}
-                        </AlertDescription>
-                      </Alert>
-                    )}
-
-                    <div className="space-y-2">
-                      <div className="text-sm font-medium">Manifest</div>
-                      <JsonViewer json={selectedArtifactDetail} />
-                    </div>
-                  </>
-                ) : (
-                  <div className="rounded border border-dashed p-6 text-sm text-muted-foreground">
-                    Select an artifact manifest to inspect its route, generation, chunks, ranges,
-                    errors, and raw manifest state.
-                  </div>
-                )}
-              </CardContent>
-            </Card>
-          </div>
+          <DocumentArtifactsPanel key={tableName} tableName={tableName || ""} />
         )}
 
         {/* Reprocess Section */}
         {currentSection === "reprocess" && (
-          <div className="grid gap-4 xl:grid-cols-[420px_1fr]">
-            <Card>
-              <CardHeader>
-                <CardTitle>Table Reprocess</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="space-y-2">
-                  <Label htmlFor="reprocess-artifact-name">Artifact name</Label>
-                  <Input
-                    id="reprocess-artifact-name"
-                    value={reprocessArtifactName}
-                    onChange={(event) => setReprocessArtifactName(event.target.value)}
-                    placeholder="document_units_v1"
-                  />
-                </div>
-                <div className="grid gap-3 md:grid-cols-2">
-                  <div className="space-y-2">
-                    <Label htmlFor="reprocess-from-key">From key</Label>
-                    <Input
-                      id="reprocess-from-key"
-                      value={reprocessFromKey}
-                      onChange={(event) => setReprocessFromKey(event.target.value)}
-                      placeholder="exclusive"
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="reprocess-to-key">To key</Label>
-                    <Input
-                      id="reprocess-to-key"
-                      value={reprocessToKey}
-                      onChange={(event) => setReprocessToKey(event.target.value)}
-                      placeholder="inclusive"
-                    />
-                  </div>
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="reprocess-limit">Pass limit</Label>
-                  <Input
-                    id="reprocess-limit"
-                    value={reprocessLimit}
-                    onChange={(event) => setReprocessLimit(event.target.value)}
-                    inputMode="numeric"
-                  />
-                </div>
-
-                {reprocessError && (
-                  <Alert variant="destructive">
-                    <AlertDescription>{reprocessError}</AlertDescription>
-                  </Alert>
-                )}
-
-                <div className="flex flex-wrap gap-2">
-                  <Button onClick={handleStartReprocessJob} disabled={reprocessLoading}>
-                    Start job
-                  </Button>
-                  <Button
-                    variant="outline"
-                    onClick={handleRunReprocessPass}
-                    disabled={reprocessLoading}
-                  >
-                    Run bounded pass
-                  </Button>
-                </div>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader>
-                <CardTitle>Reprocess Status</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                {reprocessJob ? (
-                  <>
-                    <div className="flex flex-wrap items-center gap-2">
-                      <Badge className={jobStatusClass(reprocessJob)}>{reprocessJob.phase}</Badge>
-                      <Badge>{reprocessJob.reprocess_status}</Badge>
-                      <span className="text-sm text-muted-foreground">
-                        Job {reprocessJob.job_id} for {reprocessJob.artifact_name}
-                      </span>
-                    </div>
-                    <div className="grid gap-3 md:grid-cols-5">
-                      <div>
-                        <div className="text-xs text-muted-foreground">Scanned</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessJob.scanned)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Reprocessed</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessJob.reprocessed)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Skipped</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessJob.skipped)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Failed</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessJob.failed)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Pending Shards</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessJob.pending_shards)}
-                        </div>
-                      </div>
-                    </div>
-                    <div className="flex flex-wrap gap-2">
-                      <Button
-                        variant="outline"
-                        onClick={handleAdvanceReprocessJob}
-                        disabled={
-                          reprocessLoading ||
-                          ["succeeded", "failed", "cancelled"].includes(reprocessJob.phase)
-                        }
-                      >
-                        Advance
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        onClick={handleCancelReprocessJob}
-                        disabled={
-                          reprocessLoading ||
-                          ["succeeded", "failed", "cancelled"].includes(reprocessJob.phase)
-                        }
-                      >
-                        Cancel
-                      </Button>
-                    </div>
-                    {reprocessJob.failures.length > 0 && (
-                      <JsonViewer json={{ failures: reprocessJob.failures }} />
-                    )}
-                  </>
-                ) : reprocessPass ? (
-                  <>
-                    <div className="flex flex-wrap items-center gap-2">
-                      <Badge>{reprocessPass.reprocess_status}</Badge>
-                      <span className="text-sm text-muted-foreground">
-                        Bounded pass for {reprocessPass.artifact_name}
-                      </span>
-                    </div>
-                    <div className="grid gap-3 md:grid-cols-4">
-                      <div>
-                        <div className="text-xs text-muted-foreground">Scanned</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessPass.scanned)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Reprocessed</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessPass.reprocessed)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Skipped</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessPass.skipped)}
-                        </div>
-                      </div>
-                      <div>
-                        <div className="text-xs text-muted-foreground">Failed</div>
-                        <div className="text-lg font-semibold">
-                          {formatCount(reprocessPass.failed)}
-                        </div>
-                      </div>
-                    </div>
-                    <JsonViewer json={reprocessPass} />
-                  </>
-                ) : (
-                  <div className="rounded border border-dashed p-6 text-sm text-muted-foreground">
-                    Start a durable job for table-wide repair, or run one bounded pass when you want
-                    to manually inspect cursor progress.
-                  </div>
-                )}
-              </CardContent>
-            </Card>
-          </div>
+          <TableReprocessPanel key={tableName} tableName={tableName || ""} />
         )}
 
         {/* Schema Section */}
